Refactor main bots tab into single component
